Update utils.py
Actioning Lou's comments to add try-catch for API connections. and if__main__ statement. Note: This try-catch checks if the API request connection was made successfully, it does not catch for rand_character() if the authentication access_token is not correct - not sure if we need to/ how we would capture that?